Separate Subscribable and UrlBuildable concerns
