+++ app/models/gobierto_budget_consultations/consultation.rb
@@ -3,6 +3,7 @@ require_dependency "gobierto_budget_consultations"
 module GobiertoBudgetConsultations
   class Consultation < ApplicationRecord
     include User::Subscribable
+    include GobiertoCommon::UrlBuildable
 
     belongs_to :admin, class_name: "GobiertoAdmin::Admin"
     belongs_to :site
