+++ app/models/gobierto_calendars/event.rb
@@ -13,6 +13,7 @@ module GobiertoCalendars
 
     include ActsAsParanoidAliases
     include User::Subscribable
+    include GobiertoCommon::UrlBuildable
     include GobiertoCommon::Searchable
     include GobiertoCommon::Sluggable
     include GobiertoCommon::Collectionable
