+++ app/models/gobierto_participation/contribution_container.rb
@@ -8,6 +8,7 @@ module GobiertoParticipation
 
     include ActsAsParanoidAliases
     include User::Subscribable
+    include GobiertoCommon::UrlBuildable
     include GobiertoCommon::Sluggable
     include GobiertoCommon::HasVisibilityUserLevels
