+++ app/models/gobierto_participation/process.rb
@@ -8,6 +8,7 @@ module GobiertoParticipation
 
     include ActsAsParanoidAliases
     include User::Subscribable
+    include GobiertoCommon::UrlBuildable
     include GobiertoCommon::Sluggable
     include GobiertoCommon::Searchable
     include GobiertoCommon::ActsAsCollectionContainer
